+++ packages/web-app/src/utils/library.ts
@@ -1,6 +1,13 @@
 // Library utils / Ethers for now
 import {BigNumberish, constants, ethers} from 'ethers';
 import {TFunction} from 'react-i18next';
+import {ApolloClient} from '@apollo/client';
+import {Client} from '@aragon/sdk-client';
+
+import {fetchTokenData} from 'services/prices';
+import {SupportedNetworks} from 'utils/constants';
+
+import {ActionWithdraw} from 'utils/types';
 
 export function formatUnits(amount: BigNumberish, decimals: number) {
   if (amount.toString().includes('.') || !decimals) {
@@ -63,3 +70,50 @@ export const getUserFriendlyWalletLabel = (
 export const toHex = (num: number | string) => {
   return '0x' + num.toString(16);
 };
+
+/**
+ * DecodeWithdrawToAction
+ * Note: Please rename to useDaoMetadata once the other hook as been deprecated
+ * @param data Uint8Array action data
+ * @param client SDK client, Fetched using useClient
+ * @param apolloClient Apollo client, Fetched using useApolloClient
+ * @param network network of the dao
+ * @returns Return Decoded Withdraw action
+ */
+export async function decodeWithdrawToAction(
+  data: Uint8Array | undefined,
+  client: Client | undefined,
+  apolloClient: ApolloClient<object>,
+  network: SupportedNetworks
+): Promise<ActionWithdraw | undefined> {
+  if (!client || !data) {
+    console.error('SDK client is not initialized correctly');
+    return;
+  }
+
+  const decoded = client.decoding.withdrawAction(data);
+
+  if (!decoded) {
+    console.error('Unable to decode withdraw action');
+    return;
+  }
+
+  const response = await fetchTokenData(
+    decoded?.tokenAddress || constants.AddressZero,
+    apolloClient,
+    network
+  );
+
+  return {
+    amount: Number(decoded.amount),
+    name: 'withdraw_assets',
+    to: decoded.recipientAddress,
+    tokenAddress: response?.address || (decoded?.tokenAddress as string),
+    tokenBalance: 0, // unnecessary
+    tokenImgUrl: response?.imgUrl as string,
+    tokenName: response?.name || '',
+    tokenPrice: response?.price || 0,
+    tokenSymbol: response?.symbol || '',
+    isCustomToken: false,
+  };
+}
